+++ apps/bot/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@apps/bot",
+  "version": "1.0.0",
+  "description": "Bot app",
+  "main": "dist/index.js",
+  "author": "Supabase Community",
+  "license": "MIT",
+  "private": true,
+  "scripts": {
+    "dev": "tsx watch ./src/index.ts",
+    "build": "del dist && tsc",
+    "start": "node dist/index.js",
+    "format": "prettier --write \"**/*.ts\""
+  },
+  "dependencies": {
+    "@packages/config": "workspace:*",
+    "discord.js": "^14.7.1"
+  },
+  "devDependencies": {
+    "@types/node": "^18.11.14",
+    "del-cli": "^5.0.0",
+    "prettier": "^2.8.1",
+    "tsx": "^3.12.1",
+    "typescript": "^4.9.4"
+  }
+}
