+++ apps/bot/src/commands/done.ts
