+++ apps/bot/src/commands/index.ts
