+++ apps/bot/src/commands/pong.ts
