+++ package.json
@@ -2,25 +2,12 @@
   "name": "supacommunitybot",
   "version": "1.0.0",
   "description": "A discord bot driven by the Supabase Community",
-  "main": "dist/index.js",
   "author": "Supabase Community",
   "license": "MIT",
   "private": true,
   "scripts": {
-    "dev": "tsx watch ./src/index.ts",
-    "build": "tsc",
-    "start": "node dist/index.js",
-    "format": "prettier --write \"**/*.ts\""
-  },
-  "dependencies": {
-    "discord.js": "^14.7.1",
-    "envsafe": "^2.0.3"
-  },
-  "devDependencies": {
-    "@types/node": "^18.11.14",
-    "dotenv": "^16.0.3",
-    "prettier": "^2.8.1",
-    "tsx": "^3.12.1",
-    "typescript": "^4.9.4"
+    "build": "pnpm -r build",
+    "build:packages": "pnpm -r --filter \"@packages/*\" --parallel build",
+    "dev": "pnpm build:packages && pnpm -r --parallel dev"
   }
 }
